+++ src/client.ts
@@ -1,4 +1,5 @@
 import * as prismicT from "@prismicio/types";
+import * as prismicH from "@prismicio/helpers";
 
 import { appendPredicates } from "./lib/appendPredicates";
 import { getCookie } from "./lib/getCookie";
@@ -7,46 +8,67 @@ import {
 	FetchLike,
 	Form,
 	HttpRequestLike,
-	LinkResolver,
 	Query,
 	Ref,
 	Repository,
 	RequestInitLike
 } from "./types";
 import { buildQueryURL, BuildQueryURLArgs } from "./buildQueryURL";
-import { HTTPError } from "./HTTPError";
+import { ForbiddenError, isForbiddenErrorAPIResponse } from "./ForbiddenError";
+import { ParsingError, isParsingErrorAPIResponse } from "./ParsingError";
+import { PrismicError } from "./PrismicError";
 import * as cookie from "./cookie";
 import * as predicate from "./predicate";
 
+/**
+ * A container for items in a SimpleTTLCache.
+ */
 type SimpleTTLCacheValueContainer<T = unknown> = {
-	cachedAt: Date;
-	ttl: number;
+	expiresAt: number;
 	value: T;
 };
 
+/**
+ * A simple time-to-live cache where items are only valid for a given number of milliseconds.
+ */
 interface SimpleTTLCache {
 	get<T>(key: string): T | undefined;
 	set<T>(key: string, value: T, ttl?: number): void;
 }
 
+/**
+ * Creates a simple cache where each element has a given time to live (TTL). After the TTL has ellapsed, the value is considered stale and will not be returned if requested.
+ */
 const createSimpleTTLCache = (): SimpleTTLCache => {
 	const cache = new Map<string, SimpleTTLCacheValueContainer>();
 
 	return {
+		/**
+		 * Get a value from the cache. If the item's TTL has ellapsed, `undefined` will be returned.
+		 *
+		 * @param key Key for the item.
+		 *
+		 * @return The value for the key, if a value exists.
+		 */
 		get<T>(key: string): T | undefined {
 			const cacheValue = cache.get(key);
 
 			if (cacheValue) {
-				if (
-					new Date().getTime() - cacheValue.cachedAt.getTime() <=
-					cacheValue.ttl
-				) {
+				if (new Date().getTime() < cacheValue.expiresAt) {
 					return cacheValue.value as T;
 				}
 			}
 		},
+
+		/**
+		 * Set a value in the cache for a given key and TTL.
+		 *
+		 * @param key Key to identify the item.
+		 * @param value Value for the key.
+		 * @param ttl Number of milliseconds to consider the value fresh.
+		 */
 		set<T>(key: string, value: T, ttl: number): void {
-			cache.set(key, { cachedAt: new Date(), ttl, value });
+			cache.set(key, { expiresAt: new Date().getTime() + ttl, value });
 		}
 	};
 };
@@ -163,7 +185,7 @@ type ResolvePreviewArgs = {
 	/**
 	 * A function that maps a Prismic document to a URL within your app.
 	 */
-	linkResolver: LinkResolver;
+	linkResolver: prismicH.LinkResolverFunction;
 
 	/**
 	 * A fallback URL if the Link Resolver does not return a value.
@@ -295,14 +317,19 @@ export class Client {
 			this.queryContentFromRef(options.ref);
 		}
 
-		if (options.fetch) {
+		if (typeof options.fetch === "function") {
 			this.fetchFn = options.fetch;
 		} else if (typeof globalThis.fetch === "function") {
 			this.fetchFn = globalThis.fetch;
 		} else {
 			throw new Error(
-				"A fetch implementation was not provided. In environments where fetch is not available (including Node.js), a fetch implementation must be provided via a polyfill or the `fetch` option."
+				"A valid fetch implementation was not provided. In environments where fetch is not available (including Node.js), a fetch implementation must be provided via a polyfill or the `fetch` option."
 			);
+		}
+
+		// If the global fetch function is used, we must bind it to the global scope.
+		if (this.fetchFn === globalThis.fetch) {
+			this.fetchFn = this.fetchFn.bind(globalThis);
 		}
 	}
 
@@ -503,7 +530,7 @@ export class Client {
 		params?: Partial<BuildQueryURLArgs>
 	): Promise<Query<TDocument>> {
 		return await this.get<TDocument>(
-			appendPredicates(predicate.at("document.id", ids))(params)
+			appendPredicates(predicate.in("document.id", ids))(params)
 		);
 	}
 
@@ -532,7 +559,7 @@ export class Client {
 		params?: Partial<BuildQueryURLArgs>
 	): Promise<TDocument[]> {
 		return await this.getAll<TDocument>(
-			appendPredicates(predicate.at("document.id", ids))(params)
+			appendPredicates(predicate.in("document.id", ids))(params)
 		);
 	}
 
@@ -916,7 +943,12 @@ export class Client {
 				ref: previewToken
 			});
 
-			return args.linkResolver(document);
+			// We know we have a valid field to resolve since we are using prismicH.documentToLinkField
+			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+			return prismicH.asLink(
+				prismicH.documentToLinkField(document),
+				args.linkResolver
+			)!;
 		} else {
 			return args.defaultURL;
 		}
@@ -1108,7 +1140,7 @@ export class Client {
 				if (typeof thisRefStringOrFn === "function") {
 					const res = await thisRefStringOrFn();
 
-					if (res != null) {
+					if (typeof res === "string") {
 						return res;
 					}
 				} else if (thisRefStringOrFn) {
@@ -1159,20 +1191,58 @@ export class Client {
 		params?: Partial<BuildQueryURLArgs>
 	): Promise<T> {
 		const options = this.buildRequestOptions(params);
-		const rawRes = await this.fetchFn(url, options);
+		const res = await this.fetchFn(url, options);
 
-		// We clone to response to avoid reading an already used Response object.
-		// This could happen if the user implements their own caching solution.
-		const res = rawRes.clone();
-
-		if (res.status === 200) {
+		// eslint-disable-next-line @typescript-eslint/no-explicit-any
+		let json: any;
+		try {
 			// We can assume Prismic REST API responses will have a `application/json`
-			// Content Type.
-			return await res.json();
-		} else if (rawRes.status === 401) {
-			throw new HTTPError("Invalid access token", rawRes, url, options);
-		} else {
-			throw new HTTPError(undefined, rawRes, url, options);
+			// Content Type. If not, this will throw, signaling an invalid response.
+			json = await res.json();
+		} catch {
+			throw new PrismicError(undefined, { url });
 		}
+
+		switch (res.status) {
+			// Successful
+			case 200: {
+				return json;
+			}
+
+			// Bad Request
+			// - Invalid predicate syntax
+			// - Ref not provided (ignored)
+			case 400: {
+				if (isParsingErrorAPIResponse(json)) {
+					throw new ParsingError(json.message, {
+						url,
+						response: json
+					});
+				}
+
+				break;
+			}
+
+			// Unauthorized
+			// - Missing access token for repository endpoint
+			// - Incorrect access token for repository endpoint
+			case 401:
+			// Forbidden
+			// - Missing access token for query endpoint
+			// - Incorrect access token for query endpoint
+			case 403: {
+				if (isForbiddenErrorAPIResponse(json)) {
+					throw new ForbiddenError(
+						"error" in json ? json.error : json.message,
+						{
+							url,
+							response: json
+						}
+					);
+				}
+			}
+		}
+
+		throw new PrismicError(undefined, { url, response: json });
 	}
 }